move id_obj_manager to own file
