+++ src/utils/cw_containers.hpp
@@ -1,0 +1,12 @@
+// ==================================================================
+// Author: Ashley Zhang (ayz27@cornell.edu)
+// Date:   2/15/2025
+// Description: cw custom containers includes
+// ==================================================================
+
+#ifndef CW_CONTAINERS_HPP
+#define CW_CONTAINERS_HPP
+
+#include "containers/id_obj_manager.hpp"
+
+#endif // CW_CONTAINERS_HPP
